Moved Arclight exports to header.
