+++ Source/Arclight/Arclight.cpp
@@ -8,99 +8,6 @@
 
 #include "../Stdinclude.hpp"
 
-namespace CC_SDK
-{
-    // Implemented like ArcSDK.
-    class EXPORT_ATTR ArcID
-    {
-        wchar_t *Internalstring;
-
-    public:
-        wchar_t *Get()
-        {
-            return Internalstring;
-        }
-        void Set(const wchar_t *Input)
-        {
-            size_t Length = std::wcslen(Input);
-            std::memcpy(Internalstring, Input, Length * 2);
-        }
-
-        ArcID()
-        {
-            Internalstring = new wchar_t[256]();
-        };
-        ArcID(ArcID &Parent)
-        {
-            Internalstring = new wchar_t[256]();
-            Set(Parent.Internalstring);
-        };
-        ArcID(const wchar_t *Plaintext)
-        {
-            Internalstring = new wchar_t[256]();
-            Set(Plaintext);
-        };
-        ~ArcID()
-        {
-            delete[] Internalstring;
-        };
-
-        ArcID &operator = (const ArcID &Right) noexcept
-        {
-            Set(Right.Internalstring);
-            return *this;
-        }
-        bool operator < (const ArcID &Right)
-        {
-            return wcscmp(Internalstring, Right.Internalstring) < 0;
-        }
-    };
-    class EXPORT_ATTR LobbyID
-    {
-        wchar_t *Internalstring;
-
-    public:
-        wchar_t *Get()
-        {
-            return Internalstring;
-        }
-        void Set(const wchar_t *Input)
-        {
-            size_t Length = std::wcslen(Input);
-            std::memcpy(Internalstring, Input, Length * 2);
-        }
-
-        LobbyID()
-        {
-            Internalstring = new wchar_t[256]();
-        };
-        LobbyID(LobbyID &Parent)
-        {
-            Internalstring = new wchar_t[256]();
-            Set(Parent.Internalstring);
-        };
-        LobbyID(const wchar_t *Plaintext)
-        {
-            Internalstring = new wchar_t[256]();
-            Set(Plaintext);
-        };
-        ~LobbyID()
-        {
-            delete[] Internalstring;
-        };
-
-        LobbyID &operator = (const LobbyID &Right) noexcept
-        {
-            Set(Right.Internalstring);
-            return *this;
-        }
-        bool operator < (const LobbyID &Right)
-        {
-            return wcscmp(Internalstring, Right.Internalstring) < 0;
-        }
-    };
-}
-
 // Create a request or response from a list of parameters.
 std::wstring Createmessage(uint32_t Messagetype, std::vector<std::wstring> Parameters)
 {
